delete children setter
As discussed in https://github.com/udapi/udapi-python/issues/22#issuecomment-267732183
`nodeA.children = nodesB` is problematic because it is not clear what should happen
with the existing children of nodeA:
* delete them without any warning (that's very dangerous I think)
* keep them (but then the semantics of nodeA.children = nodesB is twisted)
* raise an exception (but then such setter would be almost useless)

So let's forbid children setter (and use parent setter instead).
It is still to be discussed whether children should stay a property
or be changed to a method (as node.descendants()) which can have parameters.