better (I hope) consistency in names "text" vs "sentence"
The CoNLL-U comment was (unfortunatelly) standardized as "# text".
It seems intuitive to store it in `root.text` (already done before).
Also the writers should have parameter `print_text` and the method
to compute this parameter on the fly should be `compute_text()`
(done now).
Moreover, `compute_text()` can be called not only on the root,
but on any node where it returns the string representation
of its subtree and it would be misleading to call it "sentence".

In addition, it seems useful to have another method for root,
which returns either `root.text` (if present) or `root.compute_text()`
and has parameter `if_missing` for specifying the behavior.
How to call this method?
For now, I've selected `root.get_sentence()`
although I am not fully satisfied with the name.