allow MISC without "=" symbol
http://universaldependencies.org/format.html says about MISC
"it has to be formatted as a list that can be split on the bar character (|)"
but unlike with FEATS, there is no requirement that each list item is
a name-value pair, with an equals sign (=) separating the name from the value.

So if there is just a name with no value, let's treat the value as True.
This way the value can be distinguished from any string ("1", "Yes",...)
and it can be serialized back (without the equals sign).
A slight disadvantage is that now you cannot write `if "xy" in node.misc[name]`,
unless you are sure that `name` is not one of the names with no value.