cosmetic update of write.TextModeTrees
A visual aid so it is easier to recognize
whether a given node precedes or follows its parent.
Imagine there are many siblings
(or some of the siblings have many descendants),
so they may end up far away from their parent
and the parent cannot be easily found.
The old visualization was:
```
 ╭─╼ a
 ├─╼ b
 ...
 ├─╼ k
─┾ parent
 ├─╼ l
 ...
 ├─╼ y
 ╰─╼ z
```

Now, we have:
```
 ╭─╼ a
 ┢─╼ b
 ...
 ┢─╼ k
─┾ parent
 ┡─╼ l
 ...
 ┡─╼ y
 ╰─╼ z
```
Thanks to @danzeman for feedback.