root.sent_id returns always the same as root.address()
`root.address()` seems redundant now, but it is parallel to
`node.address()` and `bundle.address()`, so let's keep it.

Also fixing a bug with split_docs, where the tree with `newdoc`
(which is a leftover from the last document, stored in `_buffer`)
could have `bundle.bundle_id = 1` not respecting the `sent_id`
stored in the tree.