fixing a bug with doc_json and newdoc
doc_json annotations were loaded to the previous document
if multiple documents were stored in one file and loaded
with `read.Conllu split_docs=1`.
At the time of parsing a CoNLL-U tree, we don't know
whether it will be stored in a "current" document,
or whether a new document will have to be created
(because of `newdoc` annotation).

- adding a test for this bug
- fixing it by storing the data in a special temp location
  `root.json['__doc__']` and moving to the corrent document
  later on