huge speedup: "NewTreex" benchmark 40s -> 25s
The main trick is to prevent creating new lists (and memory allocations) wherever possible.
E.g. `node.children` creates a new `ListOfNodes` object with a copy of the list of children
and `node.children(add_self=True)` creates one more copy (and the previous copy is thrown away for gc).
Thus internally, we can call node._children (which currently does not guarantee sorted result),
which creates no extra list.
`node.children(add_self=True)` was changed so it creates just a single new list.
Further speedup is possible in future.

Very minor speedup is due to direct usage of attributes instead of overloaded properties, e.g. node._ord instead of node.ord.
This is not worth the effort in user blocks, but internally in the core API it makes a (small) difference in total.