CycleError for faster and better cycle detection handling
While
 logging.warning('My message: %s', node)
stringify-ies node into the message according to %s,
 raise ValueError('My message: %s', node)
does not work like this. We need our own exception class with overridden __str__.

The reason for stringifying the message only in __str__ is speed.
When users catch the exception, they may not need to print the message.
Current str(node) is relatively slow because it calls node.address(),
which calls node.root.address().