change CorefMention.__lt__, document rules for serialization
If two mentions start at the same word,
the longer must be saved first, in the new format.
However, we cannot cycle through `reversed(doc.coref_mentions)`
because that would break the ordering of closing brackets.
The easiest solution seems to be to redefine `CorefMention.__lt__`,
so that it follows the order in which mentions must be stored
in the new format.