sanity check: cannot compare nodes from different documents
When `node1` and `node2` are from different documents,
it makes no sense to use `node1.precedes(node2)`.
Sorting nodes from different docs using `precedes` could result in unexpected bugs
because with the default Python `sort` even nodes from the same document
may not be in the correct order (the relation was not transitive without the added sanity check).