minor bugfix in eval.F1
In the rare case of no matching aligned non-focused nodes,
we have `nf_common==[]` and `c==0`, so `if c == len(nf_common)`
is true in the first iteration of the while loop and we need
`find_lcs(pred_tokens[0:], gold_tokens[0:])`, not
`find_lcs(pred_tokens[1:], gold_tokens[1:])`, which would
exclude the very first token, which could be focused and matching.

In other cases, the i+1 and j+1 worked correctly
(and resulted in a tiny speed improvement because
it was guaranteed that `pred_tokens[i] != gold_tokens[j]`).