check if IDs (node.ord) form a valid sequence in *.conll and *.conllup
There are e.g. files with gaps (1,2,3,5). Failing to detect this
results in wrong parsing and even `fix_cycles=1` may result in infinite loop.

When a node is created using `root.create_child()`,
it already has `node.ord` set up correctly as the last node in the tree.

For *.conllu we could do the same, but we don't do that
because Read.Conllu is speed critical
and invalid files can be detected using `validate.py`.